Create folders to organize files and create test files
